refactor: move jaegerai handler to handler.go
